feat(autocapture): add data extraction utilities to DataExtractor class
