+++ packages/plugin-autocapture-browser/src/helpers.ts
@@ -81,60 +81,8 @@ export const createShouldTrackEvent = (
   };
 };
 
-export const isNonSensitiveString = (text: string | null) => {
-  if (text == null) {
-    return false;
-  }
-  if (typeof text === 'string') {
-    const ccRegex =
-      /^(?:(4[0-9]{12}(?:[0-9]{3})?)|(5[1-5][0-9]{14})|(6(?:011|5[0-9]{2})[0-9]{12})|(3[47][0-9]{13})|(3(?:0[0-5]|[68][0-9])[0-9]{11})|((?:2131|1800|35[0-9]{3})[0-9]{11}))$/;
-    if (ccRegex.test((text || '').replace(/[- ]/g, ''))) {
-      return false;
-    }
-    const ssnRegex = /(^\d{3}-?\d{2}-?\d{4}$)/;
-    if (ssnRegex.test(text)) {
-      return false;
-    }
-  }
-  return true;
-};
-
 export const isTextNode = (node: Node) => {
   return !!node && node.nodeType === 3;
-};
-
-export const isNonSensitiveElement = (element: Element) => {
-  /* istanbul ignore next */
-  const tag = element?.tagName?.toLowerCase?.();
-  const isContentEditable =
-    element instanceof HTMLElement ? element.getAttribute('contenteditable')?.toLowerCase() === 'true' : false;
-
-  return !SENSITIVE_TAGS.includes(tag) && !isContentEditable;
-};
-
-// Maybe this can be simplified with element.innerText, keep and manual concatenating for now, more research needed.
-export const getText = (element: Element): string => {
-  let text = '';
-  if (isNonSensitiveElement(element) && element.childNodes && element.childNodes.length) {
-    element.childNodes.forEach((child) => {
-      let childText = '';
-      if (isTextNode(child)) {
-        if (child.textContent) {
-          childText = child.textContent;
-        }
-      } else {
-        childText = getText(child as Element);
-      }
-      text += childText
-        .split(/(\s+)/)
-        .filter(isNonSensitiveString)
-        .join('')
-        .replace(/[\r\n]/g, ' ')
-        .replace(/[ ]+/g, ' ')
-        .substring(0, 255);
-    });
-  }
-  return text;
 };
 
 export const getAttributesWithPrefix = (element: Element, prefix: string): { [key: string]: string } => {
@@ -216,22 +164,6 @@ export const getClosestElement = (element: Element | null, selectors: string[]):
   }
   /* istanbul ignore next */
   return getClosestElement(element?.parentElement, selectors);
-};
-
-// Returns the element properties for the given element in Visual Labeling.
-export const getEventTagProps = (element: Element) => {
-  if (!element) {
-    return {};
-  }
-  /* istanbul ignore next */
-  const tag = element?.tagName?.toLowerCase?.();
-
-  const properties: Record<string, JSONValue> = {
-    [constants.AMPLITUDE_EVENT_PROP_ELEMENT_TAG]: tag,
-    [constants.AMPLITUDE_EVENT_PROP_ELEMENT_TEXT]: getText(element),
-    [constants.AMPLITUDE_EVENT_PROP_PAGE_URL]: window.location.href.split('?')[0],
-  };
-  return removeEmptyProperties(properties);
 };
 
 export const asyncLoadScript = (url: string) => {
